+++ container-engine-lib/lib/backend_impls/docker/docker_manager/docker_manager.go
@@ -680,19 +680,22 @@ func (manager DockerManager) WaitForExit(context context.Context, containerId st
 }
 
 /*
-GetContainerLogs
-Gets the logs for the given container as a io.ReadCloser. The caller is responsible for closing the ReadCloser!!!
+GetContainerLogs gets the logs for the given container as a io.ReadCloser. The caller is responsible for closing the ReadCloser!!!
 
 NOTE: These logs have STDOUT and STDERR multiplexed together, and the 'stdcopy' package needs to be used to
 	demultiplex them per https://github.com/moby/moby/issues/32794
 */
-func (manager DockerManager) GetContainerLogs(context context.Context, containerId string, shouldFollowLogs bool) (io.ReadCloser, error) {
+func (manager DockerManager) GetContainerLogs(
+	ctx context.Context,
+	containerId string,
+	shouldFollowLogs bool,
+) (io.ReadCloser, error) {
 	containerLogOpts := types.ContainerLogsOptions{
 		ShowStdout: true,
 		ShowStderr: true,
 		Follow:     shouldFollowLogs,
 	}
-	readCloser, err := manager.dockerClient.ContainerLogs(context, containerId, containerLogOpts)
+	readCloser, err := manager.dockerClient.ContainerLogs(ctx, containerId, containerLogOpts)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting logs for container ID '%v'", containerId)
 	}