+++ container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts/label_value_consts.go
@@ -18,11 +18,11 @@ const (
 	apiContainerContainerTypeLabelValueStr          = "api-container"
 	userServiceContainerTypeLabelValueStr           = "user-service"
 	networkingSidecarContainerTypeLabelValueStr     = "networking-sidecar"
-	moduleContainerTypeLabelValueStr                = "module"
-	filesArtifactExpanderContainerTypeLabelValueStr = "files-artifact-expander"
+	moduleContainerTypeLabelValueStr                 = "module"
+	filesArtifactsExpanderContainerTypeLabelValueStr = "files-artifacts-expander"
 
-	enclaveDataVolumeTypeLabelValueStr = "enclave-data"
-	filesArtifactExpansionVolumeTypeLabelValueStr = "files-artifact-expansion"
+	enclaveDataVolumeTypeLabelValueStr            = "enclave-data"
+	filesArtifactExpansionVolumeTypeLabelValueStr = "files-artifacts-expansion"
 
 	trueValueStr  = "true"
 	falseValueStr = "false"
@@ -45,7 +45,7 @@ var UserServiceContainerTypeDockerLabelValue = docker_label_value.MustCreateNewD
 var NetworkingSidecarContainerTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(networkingSidecarContainerTypeLabelValueStr)
 var NetworkPartitioningEnabledDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(trueValueStr)
 var NetworkPartitioningDisabledDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(falseValueStr)
-var FilesArtifactExpanderContainerTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(filesArtifactExpanderContainerTypeLabelValueStr)
+var FilesArtifactExpanderContainerTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(filesArtifactsExpanderContainerTypeLabelValueStr)
 
 var EnclaveDataVolumeTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(enclaveDataVolumeTypeLabelValueStr)
 var FilesArtifactExpansionVolumeTypeDockerLabelValue = docker_label_value.MustCreateNewDockerLabelValue(filesArtifactExpansionVolumeTypeLabelValueStr)